extracted descriptor factory from ApplicationsTreeStructure
